Fixed line bugs

+++ src/graphics/camera.cpp
@@ -29,15 +29,23 @@ vec2 Camera::position() const {
 void Camera::compose() {
     float h_width = dimensions_.x * 0.5f;
     float h_height = dimensions_.y * 0.5f;
-    float left = position_.x - h_width;
-    float top = position_.y - h_height;
-    float right = position_.x + h_width;
-    float bottom = position_.y + h_height;
+
+//    float left = position_.x - h_width;
+//    float top = position_.y - h_height;
+//    float right = position_.x + h_width;
+//    float bottom = position_.y + h_height;
+
+    float left = (int)(position_.x - h_width);
+    float top = (int)(position_.y - h_height);
+    float right = (int)(position_.x + h_width);
+    float bottom = (int)(position_.y + h_height);
 
     float sx = 2.f / (right - left);
     float sy = 2.f / (top - bottom);
     float tx = -(right + left) / (right - left);
     float ty = -(top + bottom) / (top - bottom);
+
+
     projection_ = { { sx, 0.f, 0.f },{ 0.f, sy, 0.f },{ tx, ty, 1.f } };
 }
